+++ src/Samsara/Fermat/Provider/Distribution/Poisson.php
@@ -42,13 +42,13 @@ class Poisson
     }
 
     /**
-     * @param int|DecimalInterface $k
+     * @param int|float|DecimalInterface $k
      *
      * @return ImmutableNumber
      * @throws IntegrityConstraint
      * @throws IncompatibleObjectState
      */
-    public function probabilityOfKEvents($k)
+    public function probabilityOfKEvents($k): ImmutableNumber
     {
 
         return $this->pmf($k);
@@ -56,13 +56,13 @@ class Poisson
     }
 
     /**
-     * @param $x
+     * @param int|float|DecimalInterface $x
      *
      * @return ImmutableNumber
      * @throws IntegrityConstraint
      * @throws IncompatibleObjectState
      */
-    public function cdf($x)
+    public function cdf($x): ImmutableNumber
     {
 
         $x = Numbers::makeOrDont(Numbers::IMMUTABLE, $x);
@@ -89,13 +89,13 @@ class Poisson
      * Not sure if the usage of stats_dens_pmf_poisson() is correct here because it is undocumented in
      * the C code for the PHP extension and on PHP.net
      *
-     * @param int|DecimalInterface $x
+     * @param int|float|DecimalInterface $x
      *
      * @return ImmutableNumber
      * @throws IntegrityConstraint
      * @throws IncompatibleObjectState
      */
-    public function pmf($x)
+    public function pmf($x): ImmutableNumber
     {
         $x = Numbers::makeOrDont(Numbers::IMMUTABLE, $x);
 
@@ -116,13 +116,13 @@ class Poisson
     }
 
     /**
-     * @param int|DecimalInterface $x1
-     * @param int|DecimalInterface $x2
+     * @param int|float|DecimalInterface $x1
+     * @param int|float|DecimalInterface $x2
      *
      * @return ImmutableNumber
      * @throws IntegrityConstraint
      */
-    public function rangePmf($x1, $x2)
+    public function rangePmf($x1, $x2): ImmutableNumber
     {
         $x1 = Numbers::makeOrDont(Numbers::IMMUTABLE, $x1);
         $x2 = Numbers::makeOrDont(Numbers::IMMUTABLE, $x2);
@@ -159,7 +159,7 @@ class Poisson
      * @throws IntegrityConstraint
      * @throws IncompatibleObjectState
      */
-    public function random()
+    public function random(): ImmutableNumber
     {
         if ($this->lambda->isLessThanOrEqualTo(30)) {
             return $this->knuthRandom();
@@ -179,14 +179,14 @@ class Poisson
      * @throws OptionalExit
      * @return ImmutableNumber
      */
-    public function rangeRandom($min = 0, $max = PHP_INT_MAX, $maxIterations = 20)
+    public function rangeRandom($min = 0, $max = PHP_INT_MAX, int $maxIterations = 20): ImmutableNumber
     {
         $i = 0;
 
         do {
             $randomNumber = $this->random();
             $i++;
-        } while (($randomNumber->isGreaterThanOrEqualTo($max) || $randomNumber->isLessThanOrEqualTo($min)) && $i < $maxIterations);
+        } while (($randomNumber->isGreaterThan($max) || $randomNumber->isLessThan($min)) && $i < $maxIterations);
 
         if ($randomNumber->isGreaterThan($max) || $randomNumber->isLessThan($min)) {
             throw new OptionalExit(
@@ -208,7 +208,7 @@ class Poisson
      * @throws IntegrityConstraint
      * @throws IncompatibleObjectState
      */
-    protected function methodPARandom()
+    protected function methodPARandom(): ImmutableNumber
     {
         $randFactory = new Factory();
         $generator = $randFactory->getMediumStrengthGenerator();
@@ -269,7 +269,7 @@ class Poisson
      * @return ImmutableNumber
      * @throws IntegrityConstraint
      */
-    protected function knuthRandom()
+    protected function knuthRandom(): ImmutableNumber
     {
         $randFactory = new Factory();
         /** @var ImmutableNumber $L */